+++ src/Command/LoadFixtureGroupCommand.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Basecom\FixturePlugin\Command;
 
+use Basecom\FixturePlugin\DurationCalculator;
 use Basecom\FixturePlugin\FixtureLoader;
 use Basecom\FixturePlugin\FixtureOption;
 use Symfony\Component\Console\Attribute\AsCommand;
@@ -31,6 +32,7 @@ class LoadFixtureGroupCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        $now = new \DateTimeImmutable();
         $io  = new SymfonyStyle($input, $output);
 
         /** @var string $groupNameInput */
@@ -64,7 +66,8 @@ class LoadFixtureGroupCommand extends Command
             return Command::FAILURE;
         }
 
-        $io->success('Done!');
+        $tookSeconds = DurationCalculator::calculateSeconds($now, new \DateTimeImmutable());
+        $io->success('Done! Took '.$tookSeconds.'s');
 
         return Command::SUCCESS;
     }