+++ src/Command/LoadSingleFixtureCommand.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Basecom\FixturePlugin\Command;
 
+use Basecom\FixturePlugin\DurationCalculator;
 use Basecom\FixturePlugin\FixtureLoader;
 use Basecom\FixturePlugin\FixtureOption;
 use Symfony\Component\Console\Attribute\AsCommand;
@@ -33,7 +34,8 @@ class LoadSingleFixtureCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $io = new SymfonyStyle($input, $output);
+        $now = new \DateTimeImmutable();
+        $io  = new SymfonyStyle($input, $output);
 
         /** @var string $fixtureName */
         $fixtureName      = $input->getArgument('fixtureName');
@@ -68,7 +70,8 @@ class LoadSingleFixtureCommand extends Command
             return Command::FAILURE;
         }
 
-        $io->success('Done!');
+        $tookSeconds = DurationCalculator::calculateSeconds($now, new \DateTimeImmutable());
+        $io->success('Done! Took '.$tookSeconds.'s');
 
         return Command::SUCCESS;
     }